src/api_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
google-genai
google-adk
